fix code for compilation - 3

+++ SQLite.Interop/src/contrib/extension-functions.c
@@ -144,8 +144,8 @@ int errno;
 #if defined(_WIN32)
 typedef signed char int8_t;
 typedef unsigned char uint8_t;
-typedef signed int int16_t;
-typedef unsigned int uint16_t;
+typedef signed short int int16_t;
+typedef unsigned short int uint16_t;
 typedef signed long int int32_t;
 typedef unsigned long int uint32_t;
 typedef signed long long int int64_t;
